pyproject.toml
[project]
name = "gencal-idiomas"
version = "0.1.0"
description = ""
authors = [
    {name = "juancparama",email = "jaracun@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pandas (>=2.3.1,<3.0.0)",
    "customtkinter (>=5.2.2,<6.0.0)",
    "tkcalendar (>=1.6.1,<2.0.0)",
    "pyodbc (>=5.2.0,<6.0.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "office365-rest-python-client (>=2.6.2,<3.0.0)",
    "requests-ntlm (>=1.3.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
